requirements.txt
streamlit==1.27.2
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
websocket-client==1.6.1
python-dateutil==2.8.2
pandas==2.0.3
numpy==1.24.4
SQLAlchemy==2.0.19
python-dotenv==1.0.0
plotly==5.17.0
matplotlib
websocket
